fix(api-reference): doesn't scroll to id on load (#6750)
* fix: initial scroll to id (wip)

* chore: clean up

* chore: disable logs

* docs(changeset): fix: initial scroll to id lands in random positions

* fix: outdated tests

* chore: clean up
