+++ .changeset/tame-wasps-count.md
@@ -1,0 +1,6 @@
+---
+'@scalar/api-reference': patch
+'@scalar/helpers': patch
+---
+
+fix: initial scroll to id lands in random positions
