+++ packages/api-reference/src/components/Content/Content.vue
@@ -13,6 +13,7 @@ import { Models } from '@/components/Content/Models'
 import { SectionFlare } from '@/components/SectionFlare'
 import { getXKeysFromObject } from '@/features/specification-extension'
 import { DEFAULT_INTRODUCTION_SLUG } from '@/features/traverse-schema'
+import { useFreezing } from '@/hooks/useFreezing'
 import { useNavState } from '@/hooks/useNavState'
 import { ClientSelector } from '@/v2/blocks/scalar-client-selector-block'
 import { InfoBlock } from '@/v2/blocks/scalar-info-block'
@@ -27,6 +28,8 @@ const { store, config } = defineProps<{
   config: ApiReferenceConfiguration
   store: WorkspaceStore
 }>()
+
+useFreezing()
 
 /**
  * Generate all client options so that it can be shared between the top client picker and the operations
