+++ packages/api-reference/src/components/Content/Operations/TraversedEntryContainer.vue
@@ -1,14 +1,12 @@
 <script setup lang="ts">
 import { useActiveEntities, useWorkspace } from '@scalar/api-client/store'
-import { freezeAtTop } from '@scalar/helpers/dom/freeze-at-top'
 import { getSlugUid } from '@scalar/oas-utils/transforms'
 import type { OpenAPIV3_1 } from '@scalar/openapi-types'
 import type { ApiReferenceConfiguration } from '@scalar/types'
 import type { WorkspaceStore } from '@scalar/workspace-store/client'
-import { computed, ref } from 'vue'
+import { computed } from 'vue'
 
 import { getCurrentIndex } from '@/components/Content/Operations/get-current-index'
-import { hasLazyLoaded, lazyBus } from '@/components/Lazy/lazyBus'
 import { useSidebar } from '@/features/sidebar'
 import { useNavState } from '@/hooks/useNavState'
 import type { ClientOptionGroup } from '@/v2/blocks/scalar-request-example-block/types'
@@ -20,10 +18,6 @@ const { document, config } = defineProps<{
   config: ApiReferenceConfiguration
   clientOptions: ClientOptionGroup[]
   store: WorkspaceStore
-}>()
-
-const emit = defineEmits<{
-  allEntriesLoaded: [loaded: true]
 }>()
 
 const { collections, servers } = useWorkspace()
@@ -67,63 +61,24 @@ const activeServer = computed(() => {
 })
 
 const { items } = useSidebar()
-const { hash, isIntersectionEnabled } = useNavState()
-
-/** Tries to freeze the scroll position of the element */
-const unfreeze = freezeAtTop(hash.value)
-
-/** Resume scrolling */
-const resume = () => {
-  unfreeze?.()
-  hasLazyLoaded.value = true
-  isIntersectionEnabled.value = true
-}
-
-/** IDs for all lazy elements above the current entry */
-const lazyIds = ref<Set<string>>(new Set())
+const { hash } = useNavState()
 
 /** The index of the root entry */
 const rootIndex = computed(() =>
   getCurrentIndex(hash.value, items.value.entries),
 )
-
-// Use the lazybus to handle [un]freezing elements
-lazyBus.on(({ loading, loaded, save }) => {
-  if (hasLazyLoaded.value) {
-    return
-  }
-
-  // Track the previous elements that are loading
-  if (loading && save) {
-    lazyIds.value.add(loading)
-  }
-
-  // Track which elements have loaded
-  if (loaded && save) {
-    lazyIds.value.delete(loaded)
-  }
-
-  // We are empty! Unfreeze the page
-  if (lazyIds.value.size === 0) {
-    emit('allEntriesLoaded', true)
-    setTimeout(() => resume(), 300)
-  }
-})
-
-// Resume scrolling after 5 seconds as a failsafe
-setTimeout(() => resume(), 5000)
 </script>
 
 <template>
   <div v-if="items.entries.length && activeCollection">
     <!-- Use recursive component for cleaner rendering -->
     <TraversedEntry
-      :entries="items.entries"
       :activeCollection
       :activeServer
       :clientOptions
       :config
       :document
+      :entries="items.entries"
       :rootIndex
       :store />
   </div>
